+++ industrialPasantias/src/main/java/com/industrial/pasantias/Repository/RubroRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.industrial.pasantias.Model.Rubro;
 
-public interface RubroRepository extends JpaRepository<Rubro, Long> {
+public interface RubroRepository extends JpaRepository<Rubro, Integer> {
     //
 }
